Creating a file per existent class and a folder per first class of every file
