+++ src/Rates/Rates.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Blockchain\Rates;
+
+class Rates
+{
+    public function __construct(Blockchain $blockchain)
+    {
+        $this->blockchain = $blockchain;
+    }
+
+    public function get()
+    {
+        $rates = array();
+
+        $json = $this->blockchain->get('ticker', array('format' => 'json'));
+        foreach ($json as $cur => $data) {
+            $rates[$cur] = new Ticker($cur, $data);
+        }
+
+        return $rates;
+    }
+
+    public function toBTC($amount, $symbol)
+    {
+        $params = array(
+            'currency' => $symbol,
+            'value'    => $amount,
+            'format'   => 'json'
+        );
+
+        return $this->blockchain->get('tobtc', $params);
+    }
+}
